requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# AI & LLM
openai
# Note: backboard is listed but user has been trying to uninstall it. I will leave it if required, but maybe comment it out if it's not used? User has a running command to uninstall it. Let's keep it for now but maybe I should check if it's used. The user script uses openai directly now.
# backboard 

# Data Processing
pandas
numpy

# Vector Database
chromadb

# Article Translation & Retrieval
deep_translator

# HTTP Requests
requests

# Environment Variables
python-dotenv

# Text Embeddings (Local & Free)
sentence-transformers
# torch is often installed as a dependency of sentence-transformers, but explicit instruction is good.
# However, putting the index-url in requirements.txt can be tricky for some parsers, but pip handles it.
torch --index-url https://download.pytorch.org/whl/cpu

# Utility
pyspellchecker

# Async Support
asyncio

# Database
supabase
